+++ lib/activerecord-pg-format-db-structure/deparser.rb
@@ -1,11 +1,14 @@
 # frozen_string_literal: true
 
 require "pg_query"
+require "anbt-sql-formatter/formatter"
 
 module ActiveRecordPgFormatDbStructure
   # Returns a list of SQL strings from a list of PgQuery::RawStmt.
   class Deparser
     attr_reader :source
+
+    PRETTY_INDENT_STRING = "    "
 
     def initialize(source)
       @source = source
@@ -19,21 +22,40 @@ module ActiveRecordPgFormatDbStructure
         deparse_index_stmt(raw_statement.stmt.index_stmt)
       in stmt: { alter_table_stmt: _ }
         deparse_alter_table_stmt(raw_statement.stmt.alter_table_stmt)
+      in stmt: { select_stmt: _ }
+        deparse_query_statement(raw_statement.stmt.select_stmt)
+      in stmt: { insert_stmt: _ }
+        deparse_insert_statement(raw_statement.stmt.insert_stmt)
+      in stmt: { create_table_as_stmt: _ }
+        deparse_create_table_as_stmt(raw_statement.stmt.create_table_as_stmt)
+      in stmt: { view_stmt: _ }
+        deparse_view_stmt(raw_statement.stmt.view_stmt)
       else
-        keep_original_string(raw_statement)
+        deparse_generic_raw_statement(raw_statement)
       end
     end
 
     private
 
-    def keep_original_string(raw_statement)
-      start = raw_statement.stmt_location || 0
-      stop = start + raw_statement.stmt_len
-      source[start..stop]
+    def deparse_generic_raw_statement(raw_statement)
+      generic_raw_statement_str = +"\n\n"
+      generic_raw_statement_str << PgQuery.deparse(
+        PgQuery::ParseResult.new(
+          version: PgQuery::PG_VERSION_NUM,
+          stmts: [raw_statement]
+        )
+      )
+      generic_raw_statement_str << ";"
     end
 
     def deparse_stmt(stmt)
       "\n#{PgQuery.deparse_stmt(stmt)};"
+    end
+
+    def deparse_query_statement(stmt)
+      generic_query_str = +"\n\n"
+      generic_query_str << pretty_formt_sql_string(PgQuery.deparse_stmt(stmt))
+      generic_query_str << ";"
     end
 
     def deparse_index_stmt(index_stmt)
@@ -102,6 +124,73 @@ module ActiveRecordPgFormatDbStructure
           relation: { relname: "tmp" }, table_elts: [elt]
         )
       ).gsub(/\ACREATE TABLE ONLY tmp \((.*)\)\z/, '\1')
+    end
+
+    def deparse_create_table_as_stmt(stmt)
+      placeholder_query = PgQuery.parse("SELECT placeholder").tree.stmts.first.stmt.select_stmt
+
+      create_table_as_stmt_str = +"\n\n"
+      create_table_as_stmt_str << PgQuery.deparse_stmt(
+        PgQuery::CreateTableAsStmt.new(
+          **stmt.to_h,
+          query: PgQuery::Node.from(placeholder_query)
+        )
+      )
+      create_table_as_stmt_str << ";"
+
+      query_str = +"(\n"
+      query_str << pretty_formt_sql_string(PgQuery.deparse_stmt(stmt.query.inner)).gsub(/^/, PRETTY_INDENT_STRING)
+      query_str << "\n)"
+
+      create_table_as_stmt_str[PgQuery.deparse_stmt(placeholder_query)] = query_str
+      create_table_as_stmt_str
+    end
+
+    def deparse_view_stmt(stmt)
+      placeholder_query = PgQuery.parse("SELECT placeholder").tree.stmts.first.stmt.select_stmt
+
+      view_stmt_str = +"\n\n"
+      view_stmt_str << PgQuery.deparse_stmt(
+        PgQuery::ViewStmt.new(
+          **stmt.to_h,
+          query: PgQuery::Node.from(placeholder_query)
+        )
+      )
+      view_stmt_str << ";"
+
+      query_str = +"(\n"
+      query_str << pretty_formt_sql_string(PgQuery.deparse_stmt(stmt.query.inner)).gsub(/^/, PRETTY_INDENT_STRING)
+      query_str << "\n)"
+
+      view_stmt_str[PgQuery.deparse_stmt(placeholder_query)] = query_str
+      view_stmt_str
+    end
+
+    def deparse_insert_statement(insert_stmt)
+      placeholder_query = PgQuery.parse("SELECT placeholder").tree.stmts.first.stmt.select_stmt
+
+      insert_stmt_str = +"\n\n\n"
+      insert_stmt_str << PgQuery.deparse_stmt(
+        PgQuery::InsertStmt.new(
+          **insert_stmt.to_h,
+          select_stmt: PgQuery::Node.from(placeholder_query)
+        )
+      )
+      insert_stmt_str << "\n;"
+
+      query_str = pretty_formt_sql_string(PgQuery.deparse_stmt(insert_stmt.select_stmt.inner))
+      query_str.gsub!("VALUES (", "VALUES\n (")
+
+      insert_stmt_str[PgQuery.deparse_stmt(placeholder_query)] = query_str
+      insert_stmt_str
+    end
+
+    def pretty_formt_sql_string(sql)
+      rule = AnbtSql::Rule.new
+      rule.keyword = AnbtSql::Rule::KEYWORD_UPPER_CASE
+      rule.indent_string = PRETTY_INDENT_STRING
+      formatter = AnbtSql::Formatter.new(rule)
+      formatter.format(sql)
     end
   end
 end